fix Conflictos de estilos corregidos
Se han corregido múltiples conflictos de estilos en la aplicación
