+++ frontend/src/components/CustomDesign.css
@@ -17,7 +17,6 @@
     flex-direction: column;
     align-items: center;
     justify-content: center;
-    height: 300px;
     gap: 20px;
   }
   
@@ -55,6 +54,15 @@
     align-items: center;
     gap: 20px;
     margin: 20px;
+  }
+
+  .form-input {
+    background: #D9D9D9;
+    color: #004040;
+    border: 0.1em #004040 solid;
+    border-radius: 0.25em;
+    box-sizing: border-box;
+    padding: 0.1em 0.5em;
   }
   
   input[type="button"] {
