+++ frontend/src/index.css
@@ -79,18 +79,7 @@ button:focus-visible {
 }
 /*================================================*/
 
-@media (prefers-color-scheme: light) {
-  :root {
-    color: #213547;
-    background-color: #ffffff;
-  }
-  a:hover {
-    color: #747bff;
-  }
-  button {
-    background-color: #f9f9f9;
-  }
-}
+
 
 .error-message {
   color: red;
