+++ frontend/src/pages/LoginFormPage.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import LoginForm from '../components/LoginForm';
+import LoginForm from '../components/LoginForm/LoginForm';
 
 export function LoginFormPage() {
   return (
