+++ frontend/src/pages/RegisterFormPage.jsx
@@ -1,12 +1,10 @@
 import React from 'react';
-import RegisterForm from '../components/RegisterForm';
+import RegisterForm from '../components/RegisterForm/RegisterForm';
 
 export function RegisterFormPage() {
   return (
-    <div style={{ display: 'flex', justifyContent: 'center', alignItems: 'center', height: '100vh' }}>
-      <div>
-        <RegisterForm />
-      </div>
+    <div>
+      <RegisterForm />
     </div>
   );
 }
